+++ tests/test_collection.py
@@ -88,5 +88,14 @@ class TestCollectionClass(unittest.TestCase):
 
         query.token("123TOKEN")
 
-        self.assertIn("token", query.params)
-        self.assertEqual(query.params["token"], "123TOKEN")
+        self.assertIn("Authorization", query.headers)
+        self.assertEqual(query.headers["Authorization"], "123TOKEN")
+
+    def bearer_test_token(self):
+        query = CollectionQuery()
+
+        query.bearer_token("123TOKEN")
+
+        self.assertIn("Authorization", query.headers)
+        self.assertEqual(query.headers["Authorization"], "Bearer 123TOKEN")
+