+++ tests/test_service.py
@@ -75,5 +75,14 @@ class TestServiceClass(unittest.TestCase):
 
         query.token("123TOKEN")
 
-        self.assertIn("token", query.params)
-        self.assertEqual(query.params["token"], "123TOKEN")
+        self.assertIn("Authorization", query.headers)
+        self.assertEqual(query.headers["Authorization"], "123TOKEN")
+
+    def bearer_test_token(self):
+        query = ServiceQuery()
+
+        query.bearer_token("123TOKEN")
+
+        self.assertIn("Authorization", query.headers)
+        self.assertEqual(query.headers["Authorization"], "Bearer 123TOKEN")
+