+++ CHANGELOG.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.8.0]
+### Added
+- [pull/15](https://github.com/nasa/python_cmr/pull/15): New feature added to filter by granules within the circle around lat/lon
+- [pull/12](https://github.com/nasa/python_cmr/pull/12): Added environments to module level to simplify imports to `from cmr import CMR_UAT`
+### Changed
+- Changed token url to being tokens in authorization headers.
+- Add in bearer token function for use of EDL bearer token in authorization headers.
 
 ## [0.7.0]
 ### Added
@@ -31,8 +38,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Older]
 - Prior releases of this software originated from https://github.com/jddeal/python-cmr/releases
 
-[Unreleased]: https://github.com/nasa/python_cmr/compare/v0.7.0...HEAD
-[0.7.0]: https://github.com/nasa/python_cmr/compare/v0.6.0...0.7.0
+[Unreleased]: https://github.com/nasa/python_cmr/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/nasa/python_cmr/compare/v0.7.0...v0.8.0
+[0.7.0]: https://github.com/nasa/python_cmr/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/nasa/python_cmr/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/nasa/python_cmr/compare/ef0f9e7d67ce99d342a568bd6a098c3462df16d2...v0.5.0
 [Older]: https://github.com/jddeal/python-cmr/releases